requirements.txt
asgiref==3.11.1
Django==5.2.11
django-environ==0.13.0
django-redis==6.0.0
djangorestframework==3.16.1
psycopg2-binary==2.9.11
redis==7.2.0
sqlparse==0.5.5
